requirements.txt
viam-sdk>=0.4.0
RPi.GPIO
